Added new functions
veryfastft() is a fft function that sums up frequency ranges to send to light displays. also calls displays
mainpulse() is a pulse function display for leds
range() is an audio frequency viewer for the leds
range5() is an audio frequency viewer for leds only in 5 frequency ranges
Also basic code clean up happened.